tensor saving and compatibility with keras models
there is also progress on the codependece build

+++ ha_models/cuda_kernels/kernels.cu
@@ -1,7 +1,5 @@
 #include "kernels.cuh"
 
-#include <math.h>
-
 #include <stdio.h>
 
 #define MAX(a,b) (a > b ? a : b)
@@ -73,7 +71,6 @@ __global__ void hadamardTensorMapped_kernel(TensorMap_DEVICE mapT, TensorMap_DEV
 	}
 }
 
-// documentation just for this since is more important
 /*
 * @param tTarget = The tensor pool in which the results will be stored
 * @param tSource1 = First tensor pool in the multiplication operation
@@ -197,7 +194,7 @@ __global__ void funcPassMappedSigmoid_kernel(TensorMap_DEVICE m, size_t blockSiz
 __global__ void funcPassTanh_kernel(Tensor_DEVICE t, size_t size) {
 	size_t i = threadIdx.x + blockIdx.x * blockDim.x;
 	if (i < size) {
-		t[i] = tanhf((float)t[i]);
+		t[i] = tanh((float)t[i]);
 	}
 }
 
@@ -207,14 +204,14 @@ __global__ void funcPassMappedTanh_kernel(TensorMap_DEVICE m, size_t blockSize, 
 		size_t accesPoint = i + allignOffset;
 		size_t blockId = accesPoint / blockSize;
 		size_t blockIndex = accesPoint % blockSize;
-		m[blockId][blockIndex] = tanhf((float)m[blockId][blockIndex]);
+		m[blockId][blockIndex] = tanh((float)m[blockId][blockIndex]);
 	}
 }
 
 __global__ void funcPassTanh5_kernel(Tensor_DEVICE t, size_t size) {
 	size_t i = threadIdx.x + blockIdx.x * blockDim.x;
 	if (i < size) {
-		t[i] = tanhf((float)t[i] * 5.0f);
+		t[i] = tanh((float)t[i] * 5.0f);
 	}
 }
 
@@ -224,7 +221,7 @@ __global__ void funcPassMappedTanh5_kernel(TensorMap_DEVICE m, size_t blockSize,
 		size_t accesPoint = i + allignOffset;
 		size_t blockId = accesPoint / blockSize;
 		size_t blockIndex = accesPoint % blockSize;
-		m[blockId][blockIndex] = tanhf((float)m[blockId][blockIndex] * 5.0f);
+		m[blockId][blockIndex] = tanh((float)m[blockId][blockIndex] * 5.0f);
 	}
 }
